+++ ibllib/pipes/behavior_tasks.py
@@ -8,10 +8,11 @@ from one.alf.files import session_path_parts
 from one.api import ONE
 
 from ibllib.oneibl.registration import get_lab
+from ibllib.oneibl.data_handlers import ServerDataHandler
 from ibllib.pipes import base_tasks
 from ibllib.io.raw_data_loaders import load_settings, load_bpod_fronts
 from ibllib.qc.task_extractors import TaskQCExtractor
-from ibllib.qc.task_metrics import HabituationQC, TaskQC
+from ibllib.qc.task_metrics import HabituationQC, TaskQC, update_dataset_qc
 from ibllib.io.extractors.ephys_passive import PassiveChoiceWorld
 from ibllib.io.extractors.bpod_trials import get_bpod_extractor
 from ibllib.io.extractors.ephys_fpga import FpgaTrials, FpgaTrialsHabituation, get_sync_and_chn_map
@@ -72,9 +73,7 @@ class HabituationTrialsBpod(base_tasks.BehaviourTask):
         return signature
 
     def _run(self, update=True, save=True):
-        """
-        Extracts an iblrig training session
-        """
+        """Extracts an iblrig training session."""
         trials, output_files = self.extract_behaviour(save=save)
 
         if trials is None:
@@ -296,7 +295,7 @@ class ChoiceWorldTrialsBpod(base_tasks.BehaviourTask):
         }
         return signature
 
-    def _run(self, update=True, save=True):
+    def _run(self, update=True, save=True, **kwargs):
         """Extracts an iblrig training session."""
         trials, output_files = self.extract_behaviour(save=save)
         if trials is None:
@@ -305,12 +304,23 @@ class ChoiceWorldTrialsBpod(base_tasks.BehaviourTask):
             return output_files
 
         # Run the task QC
-        self.run_qc(trials)
+        qc = self.run_qc(trials, update=update, **kwargs)
+        if update and not self.one.offline:
+            on_server = self.location == 'server' and isinstance(self.data_handler, ServerDataHandler)
+            if not on_server:
+                _logger.warning('Updating dataset QC only supported on local servers')
+            else:
+                labs = get_lab(self.session_path, self.one.alyx)
+                # registered_dsets = self.register_datasets(labs=labs)
+                datasets = self.data_handler.uploadData(output_files, self.version, labs=labs)
+                update_dataset_qc(qc, datasets, self.one)
 
         return output_files
 
     def extract_behaviour(self, **kwargs):
         self.extractor = get_bpod_extractor(self.session_path, task_collection=self.collection)
+        _logger.info('Bpod trials extractor: %s.%s',
+                     self.extractor.__module__, self.extractor.__class__.__name__)
         self.extractor.default_path = self.output_collection
         return self.extractor.extract(task_collection=self.collection, **kwargs)
 
@@ -453,12 +463,11 @@ class ChoiceWorldTrialsNidq(ChoiceWorldTrialsBpod):
         if plot_qc:
             _logger.info('Creating Trials QC plots')
             try:
-                # TODO needs to be adapted for chained protocols
                 session_id = self.one.path2eid(self.session_path)
-                plot_task = BehaviourPlots(session_id, self.session_path, one=self.one)
+                plot_task = BehaviourPlots(
+                    session_id, self.session_path, one=self.one, task_collection=self.output_collection)
                 _ = plot_task.run()
                 self.plot_tasks.append(plot_task)
-
             except Exception:
                 _logger.error('Could not create Trials QC Plot')
                 _logger.error(traceback.format_exc())
@@ -466,14 +475,11 @@ class ChoiceWorldTrialsNidq(ChoiceWorldTrialsBpod):
         return qc
 
     def _run(self, update=True, plot_qc=True, save=True):
-        dsets, out_files = self.extract_behaviour(save=save)
+        output_files = super()._run(update=update, save=save, plot_qc=plot_qc)
+        if update and not self.one.offline:
+            self._behaviour_criterion(update=update)
 
-        if not self.one or self.one.offline:
-            return out_files
-
-        self._behaviour_criterion(update=update)
-        self.run_qc(dsets, update=update, plot_qc=plot_qc)
-        return out_files
+        return output_files
 
 
 class ChoiceWorldTrialsTimeline(ChoiceWorldTrialsNidq):